Fixed the query for priceEur

+++ src/main/java/com/example/demo/service/dto/ProductSearchRequest.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class ProductSearchRequest {
-    private Integer priceEur;
+    private Double priceEur;
     private String sim;
     private String announceDate;
 }
